+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.3.0
+* \+ : Setting to display new window button even if there is another window for the tab mover
+* \+ : Path-like resolving with bookmarks for new tab, and setting to customize folders to display in new tab
+
 # 2.2.0
 * \+ : Capture page system for the new tab page
 * \+ : Restore last closed tab button for the new tab page